+++ include/vinecopulib/bicop/gaussian.hpp
@@ -36,6 +36,19 @@ private:
   // inverse hfunction
   Eigen::VectorXd hinv1_raw(const Eigen::MatrixXd& u);
 
+  // parameter-aware overloads (`parameters` is 1 x m, m in {1, n})
+  Eigen::VectorXd pdf_raw(const Eigen::MatrixXd& u,
+                          const Eigen::MatrixXd& parameters) override;
+
+  Eigen::VectorXd cdf(const Eigen::MatrixXd& u,
+                      const Eigen::MatrixXd& parameters) override;
+
+  Eigen::VectorXd hfunc1_raw(const Eigen::MatrixXd& u,
+                             const Eigen::MatrixXd& parameters) override;
+
+  Eigen::VectorXd hinv1_raw(const Eigen::MatrixXd& u,
+                            const Eigen::MatrixXd& parameters) override;
+
   Eigen::MatrixXd tau_to_parameters(const double& tau);
 
   Eigen::VectorXd get_start_parameters(const double tau);